Move the GC stack tests to their own file and use tryGetValueWithoutGC
The pre-existing stress tests in error-gc-test.test.js are too slow for
debug+ASAN runs with the default per-test timeout, so the new tests get
their own file.

In the finalizer branch, read name/message with tryGetValueWithoutGC:
plain tryGetValue resolves ropes through a path that calls
heap.reportExtraMemoryAllocated, which must not run during the GC end
phase. This matches the MustNotTriggerGC lookup in ErrorStackTrace.cpp.
